AuthenticateAccountFilterService Implemented
Found out that the fist line of these messages is the title... lol
Templated custom exceptions for now. Will need to figure out a way to utilize those in the future

Added some constants in replace of plain text, still need to figure out if I want to organize those in one big folder or spread out depending on what part of app they belong to.
Added a base controller to inherit from, naming of said controller may change

Filter service added is about final, one little quirk is it requires user to be Authorized otherwise it throws an error. May be a better way for this? Inherited BaseController requires Authorization and includes attribute for service filter so I doubt this will be a problem.